backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.2.1",
    "boto3>=1.42.63",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "resend>=2.23.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "bcrypt>=4.0.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]
